omxvideoenc: Look for OMX_BUFFERFLAG_ENDOFFRAME flag.
There is a distinct chance that an encoded buffer can exceed
the allocated buffer size. This is normally signalled via
OMX_BUFFERFLAG_ENDOFFRAME only being set on the last buffer of the
encoded packet.
Wait for that flag, accumulating data as it comes in, and pass it
on when the flag arrives.